# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Job": {
        "description": "A row in the `jobs` table. The `*_json` TEXT columns are surfaced\nas parsed objects (`params` / `progress` / `result`); build one from a\n`sqlite3.Row` with `Job.from_row(row)`, not `model_validate`.",
        "properties": {
          "cancel_requested": {
            "default": false,
            "title": "Cancel Requested",
            "type": "boolean"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "kind": {
            "$ref": "#/components/schemas/JobKind"
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params"
          },
          "progress": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "default": "pending"
          }
        },
        "required": [
          "kind"
        ],
        "title": "Job",
        "type": "object"
      },
      "JobKind": {
        "description": "Kinds of background job tracked in the `jobs` table (the Pipeline\nUI). `import` / `export` predate this and keep their own tables.",
        "enum": [
          "enrich",
          "categorize",
          "digest",
          "pull",
          "model_pull",
          "reembed"
        ],
        "title": "JobKind",
        "type": "string"
      },
      "JobStartResponse": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "integer"
          }
        },
        "required": [
          "job_id"
        ],
        "title": "JobStartResponse",
        "type": "object"
      },
      "JobStatus": {
        "description": "Shared status enum for long-running jobs (import, export, and the\n`jobs` table). `CANCELLED` only applies to `jobs`-table jobs.",
        "enum": [
          "pending",
          "running",
          "done",
          "failed",
          "cancelled"
        ],
        "title": "JobStatus",
        "type": "string"
      },
...
      "ModelPullRequest": {
        "properties": {
          "model": {
            "title": "Model",
            "type": "string"
          }
        },
        "required": [
          "model"
        ],
        "title": "ModelPullRequest",
        "type": "object"
      },
...
      "OllamaModelInfo": {
        "properties": {
          "modified_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modified At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          }
        },
        "required": [
          "name"
        ],
        "title": "OllamaModelInfo",
        "type": "object"
      },
...
      "ProviderInfo": {
        "properties": {
          "api_key_set": {
            "default": false,
            "title": "Api Key Set",
            "type": "boolean"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "kind"
        ],
        "title": "ProviderInfo",
        "type": "object"
      },
...
      "SecretUpdate": {
        "properties": {
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "auth_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auth Token"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "set_auth_token": {
            "default": false,
            "title": "Set Auth Token",
            "type": "boolean"
          }
        },
        "title": "SecretUpdate",
        "type": "object"
      },
...
      "TaskRouting": {
        "properties": {
          "model": {
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "source": {
            "enum": [
              "configured",
              "default"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "model",
          "source"
        ],
        "title": "TaskRouting",
        "type": "object"
      },
      "TaskRoutingUpdate": {
        "properties": {
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "provider_kind": {
            "anyOf": [
              {
                "enum": [
                  "ollama",
                  "openai",
                  "anthropic"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Kind"
          },
          "task": {
            "title": "Task",
            "type": "string"
          }
        },
        "required": [
          "task",
          "provider",
          "model"
        ],
        "title": "TaskRoutingUpdate",
        "type": "object"
      },
      "TaskRoutingUpdateResponse": {
        "properties": {
          "auth_token_set": {
            "title": "Auth Token Set",
            "type": "boolean"
          },
          "needs_reembed": {
            "default": false,
            "title": "Needs Reembed",
            "type": "boolean"
          },
          "ollama_models": {
            "items": {
              "$ref": "#/components/schemas/OllamaModelInfo"
            },
            "title": "Ollama Models",
            "type": "array"
          },
          "ollama_reachable": {
            "title": "Ollama Reachable",
            "type": "boolean"
          },
          "providers": {
            "items": {
              "$ref": "#/components/schemas/ProviderInfo"
            },
            "title": "Providers",
            "type": "array"
          },
          "tasks": {
            "additionalProperties": {
              "$ref": "#/components/schemas/TaskRouting"
            },
            "title": "Tasks",
            "type": "object"
          }
        },
        "required": [
          "ollama_reachable",
          "ollama_models",
          "providers",
          "tasks",
          "auth_token_set"
        ],
        "title": "TaskRoutingUpdateResponse",
        "type": "object"
      },
...
      "TestTaskRequest": {
        "properties": {
          "task": {
            "title": "Task",
            "type": "string"
          }
        },
        "required": [
          "task"
        ],
        "title": "TestTaskRequest",
        "type": "object"
      },
      "TestTaskResult": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "detail"
        ],
        "title": "TestTaskResult",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Private, local-first pipeline for saved Instagram content.",
    "title": "GramVault Atlas",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/api/models/pull": {
      "post": {
        "description": "Pull an Ollama model, tracked as a `model_pull` job \u2014 poll\n`GET /api/jobs/{job_id}` for `progress` ({status, completed, total}).",
        "operationId": "pull_model_api_models_pull_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelPullRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pull Model",
        "tags": [
          "models"
        ]
      }
    },
...
    "/api/models/secrets": {
      "put": {
        "description": "Write-only: store a provider API key, or the API auth token. Never\nreturns the stored value.",
        "operationId": "set_secret_api_models_secrets_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SecretUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Secret",
        "tags": [
          "models"
        ]
      }
    },
...
    "/api/models/tasks": {
      "put": {
        "description": "Point an AI task at a provider/model. If `provider_kind` is given,\nalso (re)registers that provider. Persisted to secrets.yaml.\n\nIf this changes an existing provider's `base_url`, its stored API key\nis cleared and must be re-entered: otherwise the next call would send\nthat key to whatever host `base_url` now points at (audit finding\nS6 \u2014 a provider's key can be silently exfiltrated by re-pointing its\nbase_url to an attacker-controlled server).",
        "operationId": "set_task_routing_api_models_tasks_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRoutingUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRoutingUpdateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Task Routing",
        "tags": [
          "models"
        ]
      }
    },
    "/api/models/test": {
      "post": {
        "description": "Run one tiny real request for a task's provider/model.",
        "operationId": "test_task_api_models_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestTaskResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Task",
        "tags": [
          "models"
        ]
      }
    },
    "/api/models/{name}": {
      "delete": {
        "operationId": "delete_model_api_models__name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Model",
        "tags": [
          "models"
        ]
      }
    },
...
  }
}